Fix tox -e coverage
In a recent PR we added some lines of code that are executed under
Python 2 only.  This caused tox -e coverage to fail.

(It caused no problems for coveralls, because they merge the coverages
of all Python versions.)

Let's use the coverage-python-version plugin so we can annotate blocks
of code that are Python version specific, so that coverage will be 100%
irrespective of which Python version you use to run the tests.