Work with messages as bytes whenever possible
This is based on the discussion in
`https://github.com/zopefoundation/Products.MailHost/issues/30`
and
`https://github.com/zopefoundation/Products.MailHost/pull/32`.

In order to allow messages with 8bit-encoding to be sent using smtplib,
they have to be prepared as bytes. Since any preparation and header
parsing and manipulation is done in `Products.MailHost`, `zope.sendmail`
only needs to store messages internally as bytes and allow them to be
passed as bytes.