Add savepoint support, so we do not fail if any code tries to create a savepoint.
Fixes issue #35.

Note that we do not need to do anything special.
There are two cases.

1. Mail sent before savepoint.
We send a mail, but delivery is delayed until the end of the transaction.
A savepoint is made.
Other changes are made.
A rollback is done to the savepoint.
The rollback method of our datamanager is called, but it does not need to do anything to restore the correct situation:
the mail is rightfully still scheduled to be mailed at the end of the transaction, if it is committed.

2. Mail sent after savepoint.
A savepoint is made.
We send a mail, but delivery is delayed until the end of the transaction.
A rollback is done to the savepoint.
Our datamanager part of the transaction is automatically discarded.